+++ backend/src/index.js
@@ -3,11 +3,11 @@ const mongoose = require("mongoose");
 const cors = require("cors");
 const { exerciseRouter } = require("../src/routes/exercise.js");
 const { workoutRouter } = require("../src/routes/workout.js");
-const { registerRouter } =require("./routes/user.js")
+const { registerRouter } =require("./routes/user.js");
 require('dotenv').config();
 const bodyParser = require("body-parser");
-const jwt = require("jsonwebtoken")
-
+const jwt = require("jsonwebtoken");
+const { verifyJWT } = require('./middleware/authMiddleware.js');
 
 const app = express();
 
@@ -23,29 +23,9 @@ app.use(cors());
 
 // app.use(logReq);
 
-const verifyJWT = (req, res, next) => {
-    const token = req.headers["x-access-token"]?.split(' ')[1]
-
-    if(token) {
-        jwt.verify(token, process.env.JWT_SECRET, (err, decoded) => {
-            if(err) return res.json({
-                isLoggedIn: false,
-                message: "Failed to Authenticate"
-            })
-            req.user = {};
-            req.user.id = decoded.id
-            req.user.username = decoded.username
-            next();
-        })
-    } else {
-        res.json({message: "Incorrect Token Given", isLoggedIn: false})
-    }
-}
-
 app.use("/user", registerRouter);
 app.use("/api/exercise", verifyJWT, exerciseRouter);
-app.use("/api/workout", workoutRouter);;
-
+app.use("/api/workout", verifyJWT, workoutRouter);
 
 async function ConnectDB() {
     await mongoose.connect(process.env.MONGO_URL, {useNewUrlParser: true, useUnifiedTopology: true});
